requirements.txt
google-genai>=0.4.0
google-cloud-storage>=2.10.0
google-cloud-discoveryengine>=0.11.0
python-dotenv>=1.0.0
python-dateutil
